+++ soap-api/src/main/java/com/soapapi/service/PublishingCompanyService.java
@@ -1,22 +1,36 @@
 package com.soapapi.service;
 
 
+import com.soapapi.exception.NotFoundException;
+import com.soapapi.exception.ValidationException;
 import com.soapapi.library.*;
+import com.soapapi.repository.BookRepository;
 import com.soapapi.repository.PublishingCompanyRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
+
+import static com.soapapi.utils.FieldsValidator.validateFields;
 
 @Service
 public class PublishingCompanyService {
 
     private final PublishingCompanyRepository publishingCompanyRepository;
+    private final BookRepository bookRepository;
 
-    public PublishingCompanyService(PublishingCompanyRepository publishingCompanyRepository) {
+    public PublishingCompanyService(PublishingCompanyRepository publishingCompanyRepository, BookRepository bookRepository) {
         this.publishingCompanyRepository = publishingCompanyRepository;
+        this.bookRepository = bookRepository;
     }
 
     public CreatePublishingCompanyResponse createPublishingCompany(CreatePublishingCompanyRequest request){
+        Map<String, Object> requiredFields = Map.of(
+                "name", request.getName()
+        );
+
+        validateFields(requiredFields);
+
         PublishingCompany pubCo = new PublishingCompany();
         pubCo.setName(request.getName());
 
@@ -29,6 +43,10 @@ public class PublishingCompanyService {
     }
 
     public GetPublishingCompanyByIdResponse getPublishingCompanyById(Long id) {
+        if(!publishingCompanyRepository.existsById(id)) {
+            throw new NotFoundException("Publishing company with id " + id + " not found.");
+        }
+
         PublishingCompany pubCo = publishingCompanyRepository.getPublishingCompanyById(id);
 
         GetPublishingCompanyByIdResponse response = new GetPublishingCompanyByIdResponse();
@@ -48,17 +66,32 @@ public class PublishingCompanyService {
 
     public UpdatePublishingCompanyResponse updatePublishingCompany(UpdatePublishingCompanyRequest request){
         UpdatePublishingCompanyResponse response = new UpdatePublishingCompanyResponse();
-        response.setSuccess(
-                publishingCompanyRepository.updatePublishingCompany(
-                        request.getId(),
-                        request.getName()
-                )
+        if(!publishingCompanyRepository.existsById(request.getId())) {
+            throw new NotFoundException("Publishing company with id " + request.getId() + " not found.");
+        }
+
+        boolean isSuccess = publishingCompanyRepository.updatePublishingCompany(
+                request.getId(),
+                request.getName()
         );
+
+        response.setSuccess(isSuccess);
+
+        if(!isSuccess){
+            throw new ValidationException("Error updating publishing company. No fields provided for update.");
+        }
 
         return response;
     }
 
     public DeletePublishingCompanyResponse deletePublishingCompany(DeletePublishingCompanyRequest request){
+        if(!publishingCompanyRepository.existsById(request.getId())) {
+            throw new NotFoundException("Publishing company with id " + request.getId() + " not found.");
+        }
+        if(bookRepository.anyDependentOnPublishingCompany(request.getId())) {
+            throw new IllegalStateException("Cannot delete publishing company with existing books.");
+        }
+
         DeletePublishingCompanyResponse response = new DeletePublishingCompanyResponse();
         response.setSuccess(publishingCompanyRepository.deletePublishingCompany(request.getId()));
